+++ Models/ResponseModels/ServiceRtDetailContent.cs
@@ -26,7 +26,7 @@ namespace Cares.Models.ResponseModels
         /// <summary>
         /// Service Rate
         /// </summary>
-        public float ServiceRate { get; set; }
+        public double ServiceRate { get; set; }
 
         /// <summary>
         /// Start Date
